Refactor: optimize PlayerDb save with PlayerRecordRef and integrate Mission/Guide systems
- Introduced PlayerRecordRef to avoid unnecessary cloning during serialization in PlayerDb::save.
- Added MissionManager and GuideManager to PlayerRecord and Player structs.
- Integrated new mission and guide command handlers in the game-server router.
- Updated session logic to use the new PlayerRecordRef for saving player state.
- Expanded scene command support including teleportation, entity lifecycle, and level script events.
- Relaxed packet validation to allow empty bodies for certain command types.

+++ servers/game-server/src/handlers/ping.rs
@@ -15,6 +15,6 @@ pub async fn on_csping(_ctx: &mut NetContext<'_>, req: CsPing) -> ScPing {
 pub async fn on_cs_flush_sync(_ctx: &mut NetContext<'_>, req: CsFlushSync) -> ScFlushSync {
     ScFlushSync {
         client_ts: req.client_ts,
-        server_ts: common::time::now_ms() as u64,
+        server_ts: common::time::now_ms(),
     }
 }
